Add the ability to filter images on upload
The session tempstore is used to handle the uploaded files providing
tools to store it in the session, get a preview url, retrieve the datas
on form validation failure.
We added the ability to provide filters when initializing the
FileTempStore, those filters will treat the file datas and return them.
Ref #601 : Revoir la gestion des logos