Ref #703 : add treasury files view
Add a view, its registration and its tests.
Treasury files need the configuration of a autonomie.ftpdir var and
should be coupled with a ftp like mechanism to allow file upload.

Treasury files should be placed in directories of the form
tresorerie/year/month/

On the same model, income statements and salarysheets should
respectivley placed in resultat and salaire directories.

The filenames should fit the following convention :
<code_compta>_year_month_week.ext

where code_compta is the accounting code of the company.