Wsgi middleware used to pipe http methods
Some http methods that are used by js modules may cause problems due to
firewalls or clients.
The Backbone provides a way to automatically pass specific http methods
as delete, patch, put through POST, passing the method as an argument in
the url. This middleware allows to transparently set the expected http
method.