Fix #838 : "partially paid" message when resulted
When an invoice is resulted, the end user get "partially paid" message
in its mail.

When a payment is registered, the new status is "paid", if the resulted
box has been checked or if the payment covers the invoice, it's set to
resulted by the state machine.
When the status changed event is sent to the registry, the initial
"paid" status is sent, here we check if the "resulted" status has been set

Note : Not all statuses are written to the model. Using only the
document.CAEStatus attribute would cause problems.