Move the menu construction to a panel
Previously the menu was set as a request attribute and was rendered in
the mako base template. All this was done using hand made solutions,
that leaked serialisation.

Now the menus are still added in the base mako template, but the
construction and parameter passing is done more cleanly (in my POV)
thanks to pyramid_layout's panels.