Change the name of the popup link
Fixes browser unpretty scroll when firing a popup.

The popup link used to use the id of the popup as link, that conduced
the browser to scroll to the top of the popup content.
Since the jquery ui dialog already handles correctly the way the dialog
is placed inside the window, we need to remove this behaviour.

Here we prepend the link's href with "popup-"