Add contribution configuration to the UI
Add an input contribution
Enhance the security through a form schema limitation :
when user hasn't manager rights, the form is reduced, previously fields
were shown but without ability to edit. Since it was possible to
submit the value whether there was no input, it was also possible to
modify key values. Now the treasury will be based on those values !!