Fix #858 : Preserve expensetype id on edition
The problem is that the links between expense lines and expense types
are broken when we edit expense types (they are all deleted and added).

Solution :
we pass the id through an hidden field to the form, and this way, we are
able to retrieve if we need to add/edit/delete the datas.