Fix #29
When doing cancelinvoice.customer = self.customer, we access the
customer object for the first time, causing a new database request.

Since the newly created cancelinvoice is pending, it's added before the
customer is retrieved causing an error (customer_id is required).

Assigning the relationship through the id avoids the database access.