Fix #41: Only highlight elements on edit/delete
When an element is newly created, a 'new_element' tag is passed to the
model.
This new_element is handled when the collectionview appends the
associated html element. If this attribute is present in the itemview's
model, we ask for highlight + scrolling, else, we only highlight.

On deletion, no scroll option is passed.