Provide panels for task rendering
Previously, we used the same html code for pdf generation and for
integration in the view_only output. The view only output contained a
full html page (with head and body tags).

Now we separate the html page and the body's content that is now in a
panel.
The html output now also handles multiple task rendering.