Use a distinct clause to ensure pagination works correctly
If we don't do that, outerjoins make the limit and count results false
(a left join b is returning 2 entries if we've got 1 "a" related to 2
"b"). It makes the pagination behave strangely (a project with many
customers can take many pages).
In order to ensure that the pagination works correctly, we add a
distinct on the id of the main model we're querying.
This way the LIMIT
clause affects only the distinct models and we get for example really 10
projects.

Conflicts:
	autonomie/views/user.py